% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_model.R
\name{coef.FishLife}
\alias{coef.FishLife}
\title{Extract path coefficients}
\usage{
\method{coef}{FishLife}(x)
}
\arguments{
\item{x}{Output from \code{\link{Fit_model}}}

\item{...}{Not used}
}
\description{
Extract path coefficients.
}
