% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Estimate_database}
\alias{Estimate_database}
\title{Database of life-history predictions}
\format{A tagged list containing data and predictions
\describe{
  \item{Y_ij}{Raw data}
  \item{Z_ik}{Taxonomy for each datum}
  ...
}}
\usage{
Estimate_database
}
\description{
Output from `Fit_model` applied to database scraped from www.FishBase.org
using `rfishbase`
}
\keyword{datasets}

