% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_ancestors.R
\name{Find_ancestors}
\alias{Find_ancestors}
\title{Find ancestors}
\usage{
Find_ancestors(child_num, ParentChild_gz = FishLife::database$ParentChild_gz)
}
\arguments{
\item{child_num}{row number of \code{ParentChild_gz} for which to find ancestors}

\item{ParentChild_gz}{Matrix representing taxonomic tree for analyzed data}
}
\value{
vector of row numbers of \code{ParentChild_gz} for ancestors (including \code{child_num})
}
\description{
Find higher taxonomic levels for a given taxon (e.g., Class and Order for a given Family)
}
