% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FishBase}
\alias{FishBase}
\title{Database of size, growth, maturity, and mortality parameters}
\format{
A tagged list containing data and predictions
\describe{
\item{N_factors}{Number of factors used for evolution in life-history model}
\item{N_obsfactors}{Number of factors used for measurent-error in life-history model}
\item{beta_gv}{Predictive mean (in transformed space) among traits for every taxon in tree}
\item{Cov_gvv}{Covariance among traits for every taxon in tree}
\item{Use_REML}{Boolean, whether REML was used for model}
\item{ParentChild_gz}{Record of taxonomic tree}
\item{ParHat}{Parameter estimates and predictions}
\item{g_i}{Associates every observation with a level of the taxonomic tree}
\item{Y_ij}{Life-history parameters from FishBase}
\item{Z_ik}{Taxonomy for each datum}
...
}
}
\usage{
FishBase
}
\description{
Output from \code{Fit_model} applied to database scraped from www.FishBase.org
using \code{rfishbase}
}
\keyword{datasets}
