% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_traits.R
\name{Get_traits}
\alias{Get_traits}
\title{Get Traits}
\usage{
Get_traits(Class = "predictive", Order = "predictive",
  Family = "predictive", Genus = "predictive", Species = "predictive",
  verbose = FALSE)
}
\arguments{
\item{Class}{Character input for taxonomic class}

\item{Order}{Character input for taxonomic class}

\item{Family}{Character input for taxonomic class}

\item{Genus}{Character input for taxonomic class}

\item{Species}{Character input for taxonomic class}

\item{verbose}{logical where TRUE prints closest match, FALSE does not}
}
\value{
a dataframe of mean trait values
}
\description{
Retrieves life history traits from FishLife
}
\details{
This function returns the mean un-logged life history traits for the closest match to the
supplied taxonomic information.
}
\examples{
life_traits <- Get_traits(Genus = "Lutjanus", Species = "campechanus")
}
