% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{Load_previous_results}
\alias{Load_previous_results}
\title{Load previous results}
\usage{
Load_previous_results(results_dir = system.file("extdata", package =
  "FishTraits"))
}
\arguments{
\item{results_dir}{Directory containing object \code{Estimate_database.RData}, archiving previous results}
}
\description{
\code{Load_previous_results} loads previous results distributed with package
}

