% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match_species.R
\name{Match_species}
\alias{Match_species}
\title{Match species}
\usage{
Match_species(genus_species = "Sebastes jordani",
  ParentChild_gz = FishLife::database$ParentChild_gz)
}
\arguments{
\item{genus_species}{Character vector for Genus_species (separated by underscore character)}

\item{ParentChild_gz}{Matrix representing taxonomic tree for analyzed data}
}
\value{
integer of row numbers of \code{ParentChild_gz} matching \code{genus_species}
}
\description{
Match \code{genus_species} to a given row of \code{ParentChild_gz}
}

