% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plot_ellipse.R
\name{Plot_ellipse}
\alias{Plot_ellipse}
\title{Plot an ellipse}
\usage{
Plot_ellipse(Cov, Mean = rep(0, 2), add = FALSE, prob = 0.95,
  xlim = log(c(0.01, 2)), ylim = xlim, logticks = c(1, 2, 5),
  ticks = c(0, 5), whichlog = "xy", main = "", xlab = "", ylab = "",
  lcol = "black", plot_lines = FALSE, ...)
}
\arguments{
\item{Cov}{2x2 matrix representing covariance}

\item{Mean}{numeric vector (length of 2) representing Empirical Bayes prediction of median}

\item{add}{Boolean whether to add ellipse to existing plot}

\item{xlim}{x-limits if \code{add=FALSE}}

\item{ylim}{y-limits if \code{add=FALSE}}

\item{logticks}{ticks to use if plotting on a log-scale}

\item{ticks}{ticks to use if plotting on a natural-scale}

\item{whichlog}{which axes are log-scale}

\item{main}{legend for each panel}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{lcol}{line color for ellipse}

\item{plot_lines}{whether to plot lines representing "major axis" and "OLS" regression parameters}

\item{probability}{for defining predictive interval}
}
\value{
integer of row numbers of \code{ParentChild_gz} matching \code{genus_species}
}
\description{
Plots an ellipse representing the bivariate predictive inteval
}

