% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_trait.R
\name{Plot_trait}
\alias{Plot_trait}
\title{Plots trait predictions}
\usage{
Plot_trait(Taxon, params = c("K", "M"),
  Cov_gjj = FishLife::database$Cov_gjj,
  Mean_gj = FishLife::database$ParHat$beta_gj,
  ParentChild_gz = FishLife::database$ParentChild_gz,
  Y_ij = FishLife::database$Y_ij, g_i = FishLife::database$g_i,
  SpeciesMatch = NULL, prob = 0.95, add = FALSE, xlim = log(c(0.01, 2)),
  ylim = xlim, partial_match = TRUE, main = "", xlab = "", ylab = "",
  lcol = "black", plot_lines = FALSE, verbose = FALSE, ticks = c(0, 5),
  logticks = c(1, 2, 5), ...)
}
\arguments{
\item{Taxon}{A taxon (matching row from \code{ParentChild_gz[,'ChildName']})}

\item{params}{character vector (length of two) of parameters to plot}

\item{Cov_gjj}{Array of estimated covariance for each taxonomic group \code{g} and trait \code{j}}

\item{Mean_gj}{Matrix of Empirical Bayes predictions of traits for each taxonomic group}

\item{ParentChild_gz}{Matrix representing taxonomic tree for analyzed data}

\item{Y_ij}{a data frame of trait values (perhaps log-scaled) with rows for records, and tagged-columns for traits}

\item{g_i}{Vector that associates every observation with a level of the taxonomic tree}

\item{SpeciesMatch}{Species for which to plot data}

\item{prob}{for defining predictive interval}

\item{add}{Boolean whether to add ellipse to existing plot}

\item{xlim}{x-limits if \code{add=FALSE}}

\item{ylim}{y-limits if \code{add=FALSE}}

\item{partial_match}{Should \code{Taxon} be an partial match or not (exact match)}

\item{main}{legend for each panel}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{lcol}{line color for ellipse}

\item{plot_lines}{whether to plot lines representing "major axis" and "OLS" regression parameters}

\item{verbose}{Boolean whether to print name matches or not}

\item{ticks}{ticks to use if plotting on a natural-scale}

\item{logticks}{ticks to use if plotting on a log-scale}

\item{...}{additional arguments passed to \code{shape::plotellipse}}
}
\description{
Plots predictive interval for a given taxon and trait-pair
}
