% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_model.R
\name{print.FishLife}
\alias{print.FishLife}
\title{Print parameter estimates}
\usage{
\method{print}{FishLife}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{Fit_model}}}

\item{...}{Not used}
}
\description{
Print parameter estimates and standard errors.
}
