% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resize_estimates.R
\name{Resize_estimates}
\alias{Resize_estimates}
\title{Resize parameter list}
\usage{
Resize_estimates(ParList, Data, N_factors, N_obsfactors)
}
\description{
Takes a list of estimated parameters, and changes the size to use as starting value for new fit
}
