% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Search_species.R
\name{Search_species}
\alias{Search_species}
\title{Search species}
\usage{
Search_species(Class = "predictive", Order = "predictive",
  Family = "predictive", Genus = "predictive", Species = "predictive",
  add_ancestors = TRUE, ParentChild_gz = Estimate_database$ParentChild_gz)
}
\arguments{
\item{Class}{Character input for taxonomic class}

\item{Order}{Character input for taxonomic class}

\item{Family}{Character input for taxonomic class}

\item{Genus}{Character input for taxonomic class}

\item{Species}{Character input for taxonomic class}

\item{add_ancestors}{Boolean whether to add ancestors for matching species or not}
}
\value{
integer of row numbers of \code{ParentChild_gz} matching \code{genus_species}
}
\description{
Match taxonomic inputs to a given row of \code{ParentChild_gz} or its closest ancestor
}

