% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAs-methods.R
\docType{methods}
\name{setAs}
\alias{setAs}
\alias{as}
\alias{as-method}
\alias{as,sem,sem-method}
\title{Convert FishLife to other package formats}
\description{
Convert output from package FishLife to package-phylobase or package-sem
}
\section{Usage}{
 \code{as(object, "phylo4d")}

 \code{as(object, "sem")}
}

\seealso{
generic \code{\link[methods]{as}}, \code{\link[sem]{sem}} from the
\code{sem} package, and \code{\link[phylobase]{phylo4d}} from the
\code{phylobase} package.
}
\keyword{methods}
