% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_ratio.R
\name{Calculate_ratio}
\alias{Calculate_ratio}
\title{Calculate ratio of two modeled traits}
\usage{
Calculate_ratio(params = c("K", "M"),
  Database = FishLife::FishBase_and_RAM, Cov_gjj = Database$Cov_gvv,
  Mean_gj = Database$beta_gv, ParentChild_gz = Database$ParentChild_gz)
}
\arguments{
\item{params}{Parameter names to use in calculation}

\item{Cov_gjj}{Array of estimated covariance for each taxonomic group \code{g} and trait \code{j}}

\item{Mean_gj}{Matrix of Empirical Bayes predictions of traits for each taxonomic group}

\item{ParentChild_gz}{Matrix representing taxonomic tree for analyzed data}
}
\value{
Matrix of exponentiated-differences (ratios) and log-standard deviations for projection
\describe{
  \item{median}{median difference for taxonomic group}
  \item{N_obsfactors}{log-standard deviation of difference for taxonomic group}
}
}
\description{
\code{Calculate_ratio} projects the covariance among traits onto a single difference (interpreted as a ratio for log-scaled traits)
}
