% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishLife.R
\docType{package}
\name{FishLife}
\alias{FishLife}
\alias{FishLife-package}
\title{Fisheries life-history database}
\description{
FishLife contains predictions of life history parameters for all described fishes,
and can also be used to replicate the multivariate random-walk model used to generate predictions,
or used to update predictions based on custom data for a single taxon.
}
\details{
Please see github README for examples \url{http://github.com/james-thorson/FishLife/}
}

