% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_model.R
\name{Fit_model}
\alias{Fit_model}
\title{Fit and predict fish traits}
\usage{
Fit_model(N_factors, N_obsfactors, Use_REML = TRUE,
  Y_ij = FishLife::database$Y_ij, Z_ik = FishLife::database$Z_ik,
  Version = "Taxon_v1_2_0", Process_cov = "Equal",
  TmbDir = system.file("executables", package = "FishLife"),
  RunDir = tempfile(pattern = "run_", tmpdir = tempdir(), fileext = "/"),
  Params = "Generate", verbose = FALSE, ...)
}
\arguments{
\item{N_factors}{Number of factors in decomposition of covariance for random-walk along evolutionary tree (0 means a diagonal but unequal covariance; negative is the sum of a factor decomposition and a diagonal-but-unequal covariance)}

\item{N_obsfactors}{Number of factors in decomposotion of observation covariance (same format as \code{N_obsfactors})}

\item{Use_REML, }{OPTIONAL boolean whether to use maximum marginal likelihood or restricted maximum likelihood (termed "REML")}

\item{Y_ij}{a data frame of trait values (perhaps log-scaled) with rows for records, and tagged-columns for traits}

\item{Z_ik}{a data frame of taxonomic classification for each row of \code{Y_ij}}

\item{Version}{TMB version number}

\item{Process_cov}{Whether process-error covariance is equal or differs multiplicatively for different taxonomic levels (Options:  "Equal" or "Unequal")}

\item{TmbDir}{Directory containing pre-compiled TMB script}

\item{RunDir}{Directory to use when compiling and running TMB script (different to avoid problems with read-write restrictions)}

\item{Params}{optional list of parameter estimates to use as starting values (Default \code{Params="Generate"} starts from random values)}

\item{verbose}{Boolean whether to print diagnostics to terminal}

\item{...}{other paramers passed to \code{TMBhelper::Optimize}}
}
\value{
Tagged list containing objects from FishLife run (first 9 slots constitute list 'Estimate_database' for archiving results)
\describe{
  \item{N_factors}{Number of factors used for evolution in life-history model}
  \item{N_obsfactors}{Number of factors used for measurent-error in life-history model}
  \item{Use_REML}{Boolean, whether REML was used for model}
  \item{Cov_gjj}{Covariance among traits for every taxon in tree}
  \item{ParentChild_gz}{Record of taxonomic tree}
  \item{ParHat}{Parameter estimates and predictions}
  \item{g_i}{Associates every observation with a level of the taxonomic tree}
  \item{Y_ij}{Raw data}
  \item{Z_ik}{Taxonomy for each datum}
  \item{Obj}{The built TMB object}
  \item{Opt}{Output from optimization}
  \item{Report}{tagged list of report-file from TMB}
  \item{ParHat_SE}{Estimated/predicted standard errors for fixed/random effects}
}
}
\description{
\code{Fit_model} estimates parameters and predicts values from a multivariate random-walk model for fish traits
}
