% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plot_taxa.R
\name{Plot_taxa}
\alias{Plot_taxa}
\title{Plots trait predictions for taxa}
\usage{
Plot_taxa(Taxa, prob = 0.95, params = matrix(c("K", "M", "Winfinity", "Loo",
  "tmax", "tm", "Lm", "Temperature"), ncol = 2, byrow = TRUE),
  Cov_gjj = Estimate_database$Cov_gjj,
  Mean_gj = Estimate_database$ParHat$beta_gj,
  ParentChild_gz = Estimate_database$ParentChild_gz,
  Y_ij = Estimate_database$Y_ij, xlim = log(c(0.01, 2)), ylim = xlim,
  ticks = c(0, 5), logticks = c(1, 2, 5), partial_match = FALSE,
  drop_pred = TRUE, mfrow = c(nrow(params), 1), legendnum = 2,
  verbose = FALSE, plot_lines = FALSE, ...)
}
\arguments{
\item{Taxa}{One or more taxonomic names (matching rows from \code{ParentChild_gz[,'ChildName']})}

\item{params}{matrix (2 columns) of parameters to plot, where each row gives parameter names for a given plot}

\item{Cov_gjj}{Array of estimated covariance for each taxonomic group \code{g} and trait \code{j}}

\item{Mean_gj}{Matrix of Empirical Bayes predictions of traits for each taxonomic group}

\item{ParentChild_gz}{Matrix representing taxonomic tree for analyzed data}

\item{Y_ij}{a data frame of trait values (perhaps log-scaled) with rows for records, and tagged-columns for traits}

\item{xlim}{x-limits if \code{add=FALSE}}

\item{ylim}{y-limits if \code{add=FALSE}}

\item{ticks}{ticks to use if plotting on a natural-scale}

\item{logticks}{ticks to use if plotting on a log-scale}

\item{partial_match}{Should \code{Taxon} be an partial match or not (exact match)}

\item{drop_pred}{Boolean whether to drop "_predictive" from legend}

\item{mfrow}{numeric vector giving number of rows and columns for panel figure}

\item{legendnum}{which panel to use for legend}

\item{verbose}{Boolean whether to print name matches or not}

\item{plot_lines}{whether to plot lines representing "major axis" and "OLS" regression parameters}
}
\description{
Plots bivariate predictive intervals for specified parameters and taxa
}

