% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_prediction.R
\name{Update_prediction}
\alias{Update_prediction}
\title{Update predictions with new data}
\usage{
Update_prediction(Taxon, Ynew_ij, partial_match = TRUE, verbose = FALSE,
  ParentChild_gz = FishLife::database$ParentChild_gz,
  Cov_gjj = FishLife::database$Cov_gjj,
  Mean_gj = FishLife::database$ParHat$beta_gj,
  obsCov_jj = FishLife::database$obsCov_jj, Version = "Update_v1_0_0",
  TmbDir = system.file("executables", package = "FishLife"),
  RunDir = tempfile(pattern = "run_", tmpdir = tempdir(), fileext = "/"))
}
\arguments{
\item{Taxon}{Boolean, whether to require a partial match to a taxon-name \code{partial_match=TRUE} , or an exact match \code{partial_match=FALSE}}

\item{Ynew_ij}{a data frame of new trait values, perhaps log-scaled, with rows for records and tagged-columns for traits}

\item{partial_match}{Boolean, whether to require a partial match to a taxon-name \code{partial_match=TRUE} , or an exact match \code{partial_match=FALSE}}

\item{verbose}{Boolean whether to print diagnostics to terminal}

\item{ParentChild_gz}{vector providing index of parent-taxon for every child-taxa}

\item{Cov_gjj}{Array of estimated covariance for each taxonomic group \code{g} and trait \code{j}}

\item{Mean_gj}{Matrix of Empirical Bayes predictions of traits for each taxonomic group}

\item{obsCov_jj}{a matrix of the observation covariance for new data}

\item{Version}{TMB version number}

\item{TmbDir}{Directory containing pre-compiled TMB script}

\item{RunDir}{Directory to use when compiling and running TMB script (different to avoid problems with read-write restrictions)}
}
\value{
Tagged list containing old and new distribution for the \code{Taxon}
\describe{
  \item{predMean_j}{Predictive mean using data from FishBase}
  \item{predCov_jj}{Predictive covariance using data from FishBase}
  \item{updateMean_j}{Updated mean after assimilating new information in \code{Ynew_ij}}
  \item{updateCov_jj}{Updated covariance after assimilating new information in \code{Ynew_ij}}
}
}
\description{
Combines the taxonomic predictions from FishBase with new user-supplied data
}
\details{
This function allows the user to update predictions using new data from one or more studies.  A study for a single parameter is entered as a new
row of \code{Ynew_ij}, where all elements are \code{NA} except for the parameter of interest.
}

